requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.6.1
pandas==2.1.3
google-cloud-storage==2.13.0
pydantic==2.5.2
sentence-transformers==3.0.1
numpy==1.26.4
